add: lifes; experience; score

+++ Assets/SoaringHorse/Scripts/Save/IPlayerProgress.cs
@@ -2,14 +2,14 @@ using System;
 
 public interface IPlayerProgress
 {
-    int Exp { get; }
-    int Score { get; }
+    float Exp { get; }
+    float Score { get; }
     int Lifes { get; }
 
     event Action Changed;
 
     void ApplyFromSave(SaveData data); // без эвента Changed
-    void AddExp(int delta);
-    void AddScore(int delta);
+    void AddExp(float delta);
+    void AddScore(float delta);
     void SetLifes(int value);
 }
